[exporterhelper] Add WithRequestQueue option to the exporter
Configuration of the queue capacity in the new exporter helper interface is based on items (spans, metric data points, or log records) instead of batches/requests as in the current exporter helper. It helps users to have better control over the queue size. It also allows putting the optional batcher sender after the queue sender, so the batcher sender doesn't need an additional queue.